+++ foode_react/src/components/Header.jsx
@@ -1,12 +1,10 @@
-import Nav from "./Nav";
-
 const Header = () => {
   return (
     <header className="">
       <h1 className="">
         5 O'Clock
       </h1>
-      <Nav />
+      
     </header>
   );
 };
