--- foode_react/src/components/Home.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import Cocktail from '../components/Cocktail'
-
-export default function Home() {
-
-    return (
-        <div>
-        <div>
-            
-            <Cocktail />
-        </div>
-        </div>
-    );
-}
