--- foode_react/src/components/Loading.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const Loading = () => {
-    return (
-     <div className="loader">
-     </div>
-    )
-   }
-   
-   export default Loading
