--- foode_react/src/components/Nav.jsx
@@ -1,15 +1,0 @@
-import { Link } from "react-router-dom";
-
-const Nav = () => {
-  return (
-    <nav className="">
-      <Link>
-        
-      </Link>
-      
-       
-    </nav>
-  );
-};
-
-export default Nav;
